Changes: - Added the following to all journal files in Stefan directory:   :file set-batch-options n n n   :file async-optimize? yes   :file cffio-options io-mode 1 - pcm was not set as material in setup journal files, this is now fixed - A steady state calculation of the temperature field is performed in case_2.jou as initialisation - Displacement added to dummy_solver.py - set_adjacent udf executed every iteration in pc_v2023R1.jou - Phase change udf's have been modified in udf_thermal.c

+++ coupling_components/solver_wrappers/fluent/pc_fluent.py
@@ -544,8 +544,9 @@ class SolverWrapperFluent(SolverWrapper):
         self.remove_dat_files(self.timestep)
 
         # delete .trn files
-        for path in glob.glob(join(self.dir_cfd, '*.trn')):
-            os.remove(path)
+        if not self.debug:
+            for path in glob.glob(join(self.dir_cfd, '*.trn')):
+                os.remove(path)
 
     def remove_dat_files(self, timestep):
         if not self.debug:
@@ -794,7 +795,7 @@ class SolverWrapperFluent(SolverWrapper):
             else:
                 dat = np.rec.fromarrays([x, y, z, id])
                 fmt = '%27.17e%27.17e%27.17e%27d'
-            tmp = f'nodes_update_timestep{self.timestep}_thread{thread_id}.dat'
+            tmp = f'nodes_update_timestep{self.timestep+1}_thread{thread_id}.dat'
             file_name = join(self.dir_cfd, tmp)
             np.savetxt(file_name, dat, fmt=fmt, header=f'{model_part.size}', comments='')
